+++ Examples/Logging/CSV files/main_rickroll.py
@@ -18,20 +18,24 @@ def get(st):
     st.append(item)
     return item
 
-servers = [
-    daf.GUILD(
-        snowflake=12345,
-        messages=[
-            daf.TextMESSAGE(None, timedelta(seconds=5), get(rolls.copy()), [12345], "edit", timedelta(seconds=5))
-        ],
-        logging=True
+accounts = [
+    daf.ACCOUNT(
+        token="SDASKDKLSADJKLSDJ",
+        is_user=False,
+        servers=[
+            daf.GUILD(
+                snowflake=12345,
+                messages=[
+                    daf.TextMESSAGE(None, timedelta(seconds=5), get(rolls.copy()), [12345], "edit", timedelta(seconds=5))
+                ],
+                logging=True
+            )
+        ]
     )
 ]
 
 
 daf.run(
-    token="OSDSJ44JNnnJNJ2NJDBWQUGHSHFAJSHDUQHFDBADVAHJVERAHGDVAHJSVDE",   # Example account token
-    is_user=False, 
-    server_list=servers,
+    accounts=accounts,
     logger=daf.LoggerCSV("History", ";")
 )