+++ src/daf/guild.py
@@ -3,8 +3,6 @@
     regarding the guild and also defines a USER class from the
     _BaseGUILD class.
 """
-from __future__ import annotations
-
 from typing import Any, Coroutine, Union, List, Optional, Dict, Callable
 from contextlib import suppress
 from typeguard import typechecked
@@ -15,7 +13,6 @@ from copy import deepcopy
 from .logging.tracing import *
 from .message import *
 
-from . import client
 from .logging import sql
 from . import misc
 from .logging import logging
@@ -72,7 +69,8 @@ class _BaseGUILD:
         "message_dict",
         "remove_after",
         "_created_at",
-        "_deleted"
+        "_deleted",
+        "parent"
     )
     
     def __init__(self,
@@ -88,6 +86,7 @@ class _BaseGUILD:
         self.remove_after = remove_after  # int - after n sends; timedelta - after amount of time; datetime - after that time
         self._deleted = False
         self._created_at = datetime.now() # The time this object was created
+        self.parent = None
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}(discord={self.apiobject})"
@@ -134,7 +133,7 @@ class _BaseGUILD:
         return (rm_after_type is timedelta and datetime.now() - self._created_at > self.remove_after or # The difference from creation time is bigger than remove_after
                 rm_after_type is datetime and datetime.now() > self.remove_after) # The current time is larger than remove_after 
 
-    def __eq__(self, other: _BaseGUILD) -> bool:
+    def __eq__(self, other: Any) -> bool:
         """
         Compares two guild objects if they're equal.
         """
@@ -157,18 +156,21 @@ class _BaseGUILD:
         """
         raise NotImplementedError
 
-    async def initialize(self, getter: Callable) -> None:
+    async def initialize(self, parent: Any, getter: Callable) -> None:
         """
         This function initializes the API related objects and then tries to initialize the MESSAGE objects.
 
         .. warning::
             This should NOT be manually called, it is called automatically after adding the message.
 
-        .. versionchanged:: v2.1
-            Merged derived classes' methods into base method to reduce code.
+        .. versionchanged:: v2.4
+            Added parent parameter to support multiple account
+            structure.
 
         Parameters
         ------------
+        parent: Any
+            The parent object. (ACCOUNT)
         getter: Callable
             Callable function or async generator used for retrieving an api object (client.get_*).
 
@@ -179,6 +181,7 @@ class _BaseGUILD:
         Other
             Raised from .add_message(message_object) method.
         """
+        self.parent = parent
         guild_id = self.snowflake
         if isinstance(self.apiobject, int):
             self.apiobject = getter(guild_id)
@@ -190,7 +193,7 @@ class _BaseGUILD:
                 try:
                     await self.add_message(message)
                 except (TypeError, ValueError) as exc:
-                    trace(f"[GUILD:] Unable to initialize message {message}, in {self}\nReason: {exc}", TraceLEVELS.WARNING)
+                    trace(f" Unable to initialize message {message}, in {self}", TraceLEVELS.WARNING, exc)
 
 
             self._messages_uninitialized.clear()
@@ -208,7 +211,7 @@ class _BaseGUILD:
             Message object to remove."""
         raise NotImplementedError
 
-    async def update(self, **kwargs):
+    async def update(self, init_options={}, **kwargs):
         """
         .. versionadded:: v2.0
 
@@ -230,17 +233,17 @@ class _BaseGUILD:
 
     @misc._async_safe("update_semaphore", 1)
     async def _advertise(self,
-                         mode: AdvertiseTaskType):
+                         type_: AdvertiseTaskType):
         """
         Main coroutine responsible for sending all the messages to this specific guild,
         it is called from the core module's advertiser task.
 
         Parameters
         --------------
-        mode: AdvertiseTaskType
+        type_: AdvertiseTaskType
             Tells which task called this method (there is one task for textual messages and one for voice like messages).
         """
-        msg_list = self.message_dict[mode]
+        msg_list = self.message_dict[type_]
         for message in msg_list[:]: # Copy the avoid issues with the list being modified while iterating (add_message/remove_message)
             # Message removal             Check due to asynchronous operations
             if message._check_state():
@@ -250,11 +253,13 @@ class _BaseGUILD:
 
             elif message._is_ready():
                 message._reset_timer()
-                message_ret = await message._send()
+                message_ctx = await message._send()
+                # Don't return anything if logging is disabled for the server or no message was sent
+                if self.logging:
+                    yield (self.generate_log_context(), message_ctx)
 
-                # Generate log (JSON or SQL)
-                if self.logging and message_ret is not None:
-                    await logging.save_log(self.generate_log_context(), message_ret)
+            yield None, None
+                    
 
     def generate_log_context(self) -> Dict[str, Union[str, int]]:
         """
@@ -351,7 +356,7 @@ class GUILD(_BaseGUILD):
         """
         self.message_dict[AdvertiseTaskType.TEXT_ISH if isinstance(message, TextMESSAGE) else AdvertiseTaskType.VOICE].remove(message)
 
-    async def initialize(self) -> None:
+    async def initialize(self, parent: Any) -> None:
         """
         This function initializes the API related objects and then tries to initialize the MESSAGE objects.
 
@@ -366,10 +371,10 @@ class GUILD(_BaseGUILD):
         Other
             Raised from .add_message(message_object) method.
         """
-        return await super().initialize(client.get_client().get_guild)
+        return await super().initialize(parent, parent.client.get_guild)
     
     @misc._async_safe("update_semaphore", 2) # Take 2 since 2 tasks share access
-    async def update(self, **kwargs):
+    async def update(self, init_options={}, **kwargs):
         """
         Used for changing the initialization parameters the object was initialized with.
 
@@ -394,8 +399,11 @@ class GUILD(_BaseGUILD):
         # Update the guild
         if "snowflake" not in kwargs:
             kwargs["snowflake"] = self.snowflake
+        
+        if len(init_options) == 0:
+            init_options = {"parent": self.parent}
 
-        await misc._update(self, **kwargs)
+        await misc._update(self, **kwargs, init_options=init_options)
         # Update messages
         for message in self.messages:
             await message.update(_init_options={"parent": self})
@@ -481,7 +489,7 @@ class USER(_BaseGUILD):
         """
         self.message_dict[AdvertiseTaskType.TEXT_ISH].remove(message)
 
-    async def initialize(self):
+    async def initialize(self, parent: Any):
         """
         This function initializes the API related objects and then tries to initialize the MESSAGE objects.
 
@@ -492,10 +500,10 @@ class USER(_BaseGUILD):
         Other
             Raised from .add_message(message_object) method.
         """
-        return await super().initialize(client.get_client().get_or_fetch_user)
+        return await super().initialize(parent, parent.client.get_or_fetch_user)
 
     @misc._async_safe("update_semaphore", 2)
-    async def update(self, **kwargs):
+    async def update(self, init_options={}, **kwargs):
         """
         .. versionadded:: v2.0
 
@@ -521,12 +529,16 @@ class USER(_BaseGUILD):
         if "snowflake" not in kwargs:
             kwargs["snowflake"] = self.snowflake
 
-        await misc._update(self, **kwargs)
+        if len(init_options) == 0:
+            init_options = {"parent": self.parent}
+
+        await misc._update(self, init_options=init_options, **kwargs)
+
         # Update messages
         for message in self.messages:
             await message.update(_init_options={"parent" : self})
 
-@misc.doc_category("Automatic generation")
+@misc.doc_category("Auto objects")
 class AutoGUILD:
     """
     .. versionadded:: v2.3
@@ -592,7 +604,8 @@ class AutoGUILD:
         "last_scan",
         "_created_at",
         "_deleted",
-        "_safe_sem"
+        "_safe_sem",
+        "parent"
     )
 
     @typechecked
@@ -613,6 +626,7 @@ class AutoGUILD:
         self.last_scan = 0
         self._deleted = False
         self._created_at = datetime.now()
+        self.parent = None
         misc._write_attr_once(self, "_safe_sem", asyncio.Semaphore(2))
 
     @property
@@ -656,10 +670,10 @@ class AutoGUILD:
         return (rm_after_type is timedelta and datetime.now() - self._created_at > self.remove_after or # The difference from creation time is bigger than remove_after
                 rm_after_type is datetime and datetime.now() > self.remove_after) # The current time is larger than remove_after 
 
-    async def initialize(self):
+    async def initialize(self, parent: Any):
         "Initializes asynchronous elements."
         # Nothing is needed to be done here since everything is obtained from API wrapper cache.
-        pass 
+        self.parent = parent
     
     async def add_message(self, message: BaseMESSAGE):
         """
@@ -703,7 +717,7 @@ class AutoGUILD:
         Coroutine that finds new guilds from the
         API wrapper.
         """
-        dcl = client.get_client()
+        dcl = self.parent.client
         stamp = datetime.now().timestamp()
         if stamp - self.last_scan > self.interval:
             self.last_scan = stamp
@@ -716,10 +730,10 @@ class AutoGUILD:
                         new_guild = GUILD(snowflake=dcgld,
                                                 messages=deepcopy(self.messages),
                                                 logging=self.logging)
-                        await new_guild.initialize()
+                        await new_guild.initialize(parent=self.parent)
                         self.cache[dcgld] = new_guild
                     except Exception as exc:
-                        trace(f"[AutoGUILD:] Unable to add new object.\nReason: {exc}",TraceLEVELS.WARNING)
+                        trace(f" Unable to add new object.",TraceLEVELS.WARNING, exc)
 
     @misc._async_safe("_safe_sem", 1)
     async def _advertise(self, type_: AdvertiseTaskType):
@@ -734,10 +748,11 @@ class AutoGUILD:
         """
         await self._process()
         for g in self.guilds:
-            await g._advertise(type_)
+            async for context in g._advertise(type_):
+                yield context # guild_ctx, message_ctx
 
     @misc._async_safe("_safe_sem", 2)
-    async def update(self, **kwargs):
+    async def update(self, init_options={}, **kwargs):
         """
         Updates the object with new initialization parameters.
 
@@ -747,4 +762,7 @@ class AutoGUILD:
         if "interval" not in kwargs:
             kwargs["interval"] = timedelta(seconds=self.interval)
 
-        return await misc._update(self, **kwargs)
+        if len(init_options) == 0:
+            init_options = {"parent": self.parent}
+
+        return await misc._update(self, init_options=init_options, **kwargs)