+++ src/daf/misc.py
@@ -5,12 +5,10 @@
 from typing import Coroutine, Callable, Any, Dict, Optional, Union
 from asyncio import Semaphore
 from functools import wraps
-from inspect import getfullargspec
+from inspect import getfullargspec, isasyncgenfunction
 from copy import copy
 from typeguard import typechecked
-
-import signal
-import sys
+from os import environ
 
 ###############################
 # Safe access functions
@@ -97,33 +95,6 @@ async def _update(obj: Any, *, init_options: dict = {}, **kwargs):
 ###########################
 # Decorators
 ###########################
-def sigint_handler(signum, frame):
-    pass
-
-def _async_cancellation_safe(func: Callable):
-    """
-    Decorator that wraps coroutine and prevents the interrupt signal.
-
-    Parameters
-    ------------
-    func: Callable
-        Function which returns a coroutine.
-    """
-    @wraps(func)
-    async def wrapper(*args, **kwargs):
-        old_handler = signal.signal(signal.SIGINT, sigint_handler)
-        return_ = await func(*args, **kwargs)
-        if old_handler != sigint_handler:
-            # This can happen since we have an await statement.
-            # If we create 2 tasks Task1 and Task2 and Task1 finishes before Task2,
-            # then Task2 would eventually set the signal handler back to sigint_handler
-            # instead to the original system handler
-            signal.signal(signal.SIGINT, old_handler)
-        
-        return return_
-
-    return wrapper
-
 @typechecked
 def _async_safe(semaphore: Union[str, Semaphore], amount: Optional[int]=1) -> Callable:
     """
@@ -154,7 +125,19 @@ def _async_safe(semaphore: Union[str, Semaphore], amount: Optional[int]=1) -> Ca
         Decorator that returns a method wrapper Coroutine that utilizes a
         asyncio semaphore to assure safe asynchronous operations.
         """
-        if isinstance(semaphore, str):
+        if isasyncgenfunction(coroutine):
+            async def wrapper(self, *args, **kwargs):
+                sem: Semaphore = getattr(self, semaphore)
+                for i in range(amount):
+                    await sem.acquire()
+                try:
+                    async for r in coroutine(self, *args, **kwargs):
+                        yield r # yield from not allowed inside async functions
+                finally:
+                    for i in range(amount):
+                        sem.release()
+
+        else:
             async def wrapper(self, *args, **kwargs):
                 sem: Semaphore = getattr(self, semaphore)
                 for i in range(amount):
@@ -166,17 +149,6 @@ def _async_safe(semaphore: Union[str, Semaphore], amount: Optional[int]=1) -> Ca
                         sem.release()
 
                 return result
-        else:
-            async def wrapper(*args, **kwargs):
-                for i in range(amount):
-                    await semaphore.acquire()
-                try:
-                    result = await coroutine(*args, **kwargs)
-                finally:
-                    for i in range(amount):
-                        semaphore.release()
-
-                return result
 
         return wraps(coroutine)(wrapper)
 
@@ -185,7 +157,7 @@ def _async_safe(semaphore: Union[str, Semaphore], amount: Optional[int]=1) -> Ca
 
 
 # Documentation
-DOCUMENTATION_MODE = "DOCUMENTATION" in sys.argv
+DOCUMENTATION_MODE = bool(environ.get("DOCUMENTATION", False))
 if DOCUMENTATION_MODE:
     doc_titles: Dict[str, list] = {}